+++ eggvance/src/registers/blendfade.cpp
@@ -2,8 +2,8 @@
 
 #include <algorithm>
 
+#include "common/bits.h"
 #include "common/macros.h"
-#include "common/utility.h"
 
 void BlendFade::reset()
 {
@@ -15,5 +15,5 @@ void BlendFade::write(int index, u8 byte)
     EGG_ASSERT(index <= 1, "Invalid index");
 
     if (index == 0)
-        evy = std::min(16, bits<0, 5>(byte));
+        evy = std::min(16, static_cast<int>(bits<0, 5>(byte)));
 }
